salama-backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.23,<1.5.0
psycopg2-binary>=2.9.1,<3.0.0
pydantic>=1.8.2,<2.0.0
python-dotenv>=0.19.0,<0.20.0
opencv-python>=4.8.0,<4.9.0
numpy>=1.24.0,<1.25.0
